+++ cuda/PV/patPV/src/GetSeeds.cu
@@ -5,10 +5,10 @@ __device__ float zCloseBeam(KalmanVeloState track, const PatPV::XYZPoint& beamsp
 {
 
   PatPV::XYZPoint tpoint(track.x, track.y, track.z);
-  PatPV::XYZPoint tdir(track.tx, track.ty, 1.);
+  PatPV::XYZPoint tdir(track.tx, track.ty, 1.f);
 
-  float wx = (1. + tdir.x * tdir.x) / track.c00;
-  float wy = (1. + tdir.y * tdir.y) / track.c11;
+  float wx = (1.f + tdir.x * tdir.x) / track.c00;
+  float wy = (1.f + tdir.y * tdir.y) / track.c11;
 
   float x0 = tpoint.x - tpoint.z * tdir.x - beamspot.x;
   float y0 = tpoint.y - tpoint.z * tdir.y - beamspot.y;
@@ -19,7 +19,7 @@ __device__ float zCloseBeam(KalmanVeloState track, const PatPV::XYZPoint& beamsp
   float yb = tpoint.y + tdir.y * (zAtBeam - tpoint.z) - beamspot.y;
   float r2AtBeam = xb * xb + yb * yb;
 
-  return r2AtBeam < 0.5 * 0.5 ? zAtBeam : 10e8;
+  return r2AtBeam < 0.5f * 0.5f ? zAtBeam : 10e8f;
 }
 
 __device__ void errorForPVSeedFinding(float tx, float ty, float& sigz2)
@@ -27,20 +27,20 @@ __device__ void errorForPVSeedFinding(float tx, float ty, float& sigz2)
 
   // the seeding results depend weakly on this eror parametrization
 
-  float pMean = 3000.; // unit: MeV
+  float pMean = 3000.f; // unit: MeV
 
   float tanTheta2 = tx * tx + ty * ty;
-  float sinTheta2 = tanTheta2 / (1. + tanTheta2);
+  float sinTheta2 = tanTheta2 / (1.f + tanTheta2);
 
   // assume that first hit in VD at 8 mm
-  float distr = 8.; // unit: mm
+  float distr = 8.f; // unit: mm
   float dist2 = distr * distr / sinTheta2;
   float sigma_ms2 = PatPV::mcu_scatCons * PatPV::mcu_scatCons * dist2 / (pMean * pMean);
-  float fslope2 = 0.0005 * 0.0005;
+  float fslope2 = 0.0005f * 0.0005f;
   float sigma_slope2 = fslope2 * dist2;
 
   sigz2 = (sigma_ms2 + sigma_slope2) / sinTheta2;
-  if (sigz2 == 0) sigz2 = 100.;
+  if (sigz2 == 0) sigz2 = 100.f;
 }
 
 __global__ void get_seeds(
@@ -76,7 +76,7 @@ __global__ void get_seeds(
     zclu = zCloseBeam(trk, beamspot);
     errorForPVSeedFinding(trk.tx, trk.ty, sigsq);
 
-    if (fabs(zclu) > 2000.) continue;
+    if (fabsf(zclu) > 2000.f) continue;
     PatPV::vtxCluster clu;
     clu.z = zclu;
     clu.sigsq = sigsq;
@@ -177,13 +177,13 @@ __device__ int find_clusters(PatPV::vtxCluster* vclus, float* zclusters, int num
 
     int n_tracks_close = 0;
     for (int i = 0; i < number_of_clusters; i++)
-      if (fabs(vclus[i].z - pvclus[index].z) < PatPV::mcu_dzCloseTracksInCluster) n_tracks_close++;
+      if (fabsf(vclus[i].z - pvclus[index].z) < PatPV::mcu_dzCloseTracksInCluster) n_tracks_close++;
 
     float dist_to_closest = 1000000.;
     if (return_number_of_clusters > 1) {
       for (int index2 = 0; index2 < return_number_of_clusters; index2++) {
-        if (index != index2 && (fabs(pvclus[index2].z - pvclus[index].z) < dist_to_closest))
-          dist_to_closest = fabs(pvclus[index2].z - pvclus[index].z);
+        if (index != index2 && (fabsf(pvclus[index2].z - pvclus[index].z) < dist_to_closest))
+          dist_to_closest = fabsf(pvclus[index2].z - pvclus[index].z);
       }
     }
 
@@ -192,7 +192,7 @@ __device__ int find_clusters(PatPV::vtxCluster* vclus, float* zclusters, int num
     bool igood = false;
     int ntracks = pvclus[index].ntracks;
     if (ntracks >= PatPV::mcu_minClusterMult) {
-      if (dist_to_closest > 10. && rat < 0.95) igood = true;
+      if (dist_to_closest > 10.f && rat < 0.95f) igood = true;
       if (ntracks >= PatPV::mcu_highMult && rat < PatPV::mcu_ratioSig2HighMult) igood = true;
       if (ntracks < PatPV::mcu_highMult && rat < PatPV::mcu_ratioSig2LowMult) igood = true;
     }
